Backend/requirements.txt
asttokens==2.2.1
backcall==0.2.0
backports.lzma==0.0.14
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.5
cmake==3.27.0
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
filelock==3.12.2
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
fonttools==4.41.1
greenlet==2.0.2
gunicorn==21.0.1
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipython==8.12.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
kiwisolver==1.4.4
lazy_loader==0.3
lit==16.0.6
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.0.74
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pylzma==0.5.0
pyodbc==4.0.39
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
PyWavelets==1.4.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.18
stack-data==0.6.2
sympy==1.12
tifffile==2023.7.10
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
Werkzeug==2.3.6
zipp==3.16.2
